requirements.txt
Django==5.2.10
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.9.0
pytz==2025.2
gunicorn==23.0.0
whitenoise>=6.6.0
PyJWT==2.9.0
cryptography==42.0.8
pytest==9.0.2
pytest-django==4.10.0
pytest-cov==6.1.0
